Remove updated() handler
The point of this handler, combined with the mounted() handler, was to ensure that the correct completion state would be set upon navigating to the stage. As it turns out, however, simply using the mounted() handler is enough for that, and thus the updated() handler was redundant, and thus could be removed.