Make stages work better
The 'stage as integer' solution works great until stage 3, which only happens if the character's discipline is a spellcaster. This solution makes it easier to deal with that case by relegating the checking to code and just doing a simple inline stage check, instead of doing the `stage == (spellcaster ? x : y)` check in the `v-if` directive.