Fix error
If `decorate()` is called before `char.race` and `char.discipline` are set it throws an error as some later stages assume the existence of those properties. That is normally fine as wizard stage 0 of a new character is the only time that either of them won't be set. Thus waiting until a later stage to call `decorate()` solves the issue.